backend/requirements.txt
# Web Framework & API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2

# AI & NLP Engine
openai==1.12.0
langchain==0.1.7
spacy==3.7.2
scispaCy==0.5.3

# Security & Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.6.1